maint: declare several "int" locals to be of type size_t
* sed/compile.c (compile_program): Using "int" was wrong: not only
too narrow, but would also evoke warnings from gcc when comparing
with variables of unsigned type.

+++ sed/compile.c
@@ -1286,7 +1286,7 @@ compile_program(vector)
 
             if (mb_cur_max > 1)
               {
-                int i, j, idx, src_char_num;
+                size_t i, j, idx, src_char_num;
                 size_t *src_lens = MALLOC(len, size_t);
                 char **trans_pairs;
                 size_t mbclen;
